add db conn pool - seperate cogs
